move rollup config to a package
 🐿 v2.8.0
