+++ components/x-teaser/package.json
@@ -19,6 +19,7 @@
   },
   "devDependencies": {
     "@financial-times/x-engine": "^1.0.0-1",
+    "@financial-times/x-rollup": "^1.0.0",
     "rollup": "^0.57.1",
     "rollup-plugin-buble": "^0.19.2",
     "rollup-plugin-commonjs": "^9.1.0"
